Reorganize bitblaster code.
